reorder in shared components
